pulled out some reporting stuff into its own file
