+++ package.json
@@ -26,7 +26,9 @@
     "typescript-eslint": "^8.23.0",
     "vitest": "^3.0.2"
   },
-  "scripts": {
-    "run:one:pdm": "cross-env TEST_FILTER=000-fundamentals MODELS=gpt-4.1 BRAINTRUST_NO_SEND_LOGS=1 BRAINTRUST_LOCAL_RESULTS=local_results.jsonl BRAINTRUST_DISABLE_PROXY=1 pdm run python -m runner.eval_convex_coding"
+  "scripts": {   
+    "local:run:one": "cross-env TEST_FILTER=000-fundamentals/000 npm run local:run",
+    "local:run": "cross-env MODELS=gpt-4.1 BRAINTRUST_NO_SEND_LOGS=1 BRAINTRUST_LOCAL_RESULTS=local_results.jsonl BRAINTRUST_DISABLE_PROXY=1 npm run run:evals",
+    "run:evals": "pdm run python -m runner.eval_convex_coding"
   }
 }
