+++ runner/eval_convex_coding.py
@@ -1,8 +1,12 @@
-from braintrust import Eval, init_logger, Reporter
-from braintrust.framework import report_failures, EvalResultWithSummary
+from braintrust import Eval, init_logger
+from braintrust.framework import EvalResultWithSummary
 from runner.models import MODELS_BY_NAME, ModelTemplate, ModelProvider
 from runner.models.model_codegen import Model
 from runner.scorer import convex_scorer, walk_answer
+from runner.reporting import (
+    convex_reporter,
+    file_reporter,
+)
 import tempfile
 from dotenv import load_dotenv
 import os
@@ -36,164 +40,8 @@ CONVEX_EVAL_ENDPOINT = os.getenv("CONVEX_EVAL_ENDPOINT")
 CONVEX_AUTH_TOKEN = os.getenv("CONVEX_AUTH_TOKEN")
 
 
-def report_eval(evaluator, result: EvalResultWithSummary, verbose, jsonl):
-    results = result.results
-    summary = result.summary
-
-    failing_results = [x for x in results if x.error]
-    if len(failing_results) > 0:
-        report_failures(evaluator, failing_results, verbose=verbose, jsonl=jsonl)
-    else:
-        num_tests = {}
-        scores = {}
-        total_score = 0
-        total_num_tests = 0
-        for eval in results:
-            if eval.metadata["category"] not in num_tests:
-                num_tests[eval.metadata["category"]] = 0
-                scores[eval.metadata["category"]] = 0
-            num_tests[eval.metadata["category"]] += 1
-            scores[eval.metadata["category"]] += eval.scores["Tests pass"]
-            total_num_tests += 1
-            total_score += eval.scores["Tests pass"]
-
-        # Post the scores to the Convex endpoint
-        if eval.metadata.get("model"):
-            try:
-                model_name = eval.metadata["model_name"]
-                # Calculate the average score for each category
-                category_scores = {
-                    category: scores[category] / num_tests[category] for category in num_tests
-                }
-                combined_score = total_score / total_num_tests
-                post_scores_to_convex(model_name, category_scores, combined_score)
-            except Exception as e:
-                print(f"Error posting scores to Convex: {e}")
-
-        # Pretty console summary
-        model_name = result.summary.project_name if hasattr(result, "summary") else None
-        overall_rate = (total_score / total_num_tests) if total_num_tests > 0 else 0
-        print()
-        print("=== Eval Summary ===")
-        print(f"Model: {eval.metadata.get('model_name', 'unknown')}")
-        print(f"Overall: {overall_rate:.2%} ({total_num_tests} tests)")
-        for category in sorted(num_tests.keys()):
-            rate = scores[category] / num_tests[category]
-            print(f"- {category}: {rate:.2%} ({num_tests[category]} tests)")
-
-        # Point to local results file when running locally
-        if os.getenv("BRAINTRUST_NO_SEND_LOGS") == "1" or os.getenv("BRAINTRUST_API_KEY") is None:
-            print(f"Results written to: {OUTPUT_RESULTS_FILE}")
-
-        # Preserve JSONL output mode if requested
-        if jsonl:
-            print(json.dumps(summary.as_dict()))
-
-    return len(failing_results) == 0
-
-
-def post_scores_to_convex(model_name, category_scores, total_score):
-    """
-    Post the evaluation scores to the Convex /updateScores endpoint.
-
-    Args:
-        model_name (str): The name of the model
-        category_scores (dict): Dictionary mapping category names to scores
-        total_score (float): The total score for the model
-    """
-    payload = {"model": model_name, "scores": category_scores, "totalScore": total_score}
-
-    if CONVEX_EVAL_ENDPOINT is not None and CONVEX_AUTH_TOKEN is not None:
-        try:
-            response = requests.post(
-                CONVEX_EVAL_ENDPOINT,
-                json=payload,
-                headers={
-                    "Content-Type": "application/json",
-                    "Authorization": f"Bearer {CONVEX_AUTH_TOKEN}",
-                },
-            )
-
-            if response.status_code == 200:
-                print(f"Successfully posted scores for model {model_name} to Convex")
-            else:
-                print(f"Failed to post scores: HTTP {response.status_code}")
-                print(f"Response: {response.text}")
-        except Exception as e:
-            print(f"Error posting scores to Convex: {str(e)}")
-
-
-def report_run(eval_reports, verbose, jsonl):
-    return all(x for x in eval_reports)
-
-
-convex_reporter = Reporter(
-    name="convex reporter",
-    report_eval=report_eval,
-    report_run=report_run,
-)
-
-# Local file reporter: writes summaries to JSONL without uploading anywhere
-OUTPUT_RESULTS_FILE = os.getenv("BRAINTRUST_LOCAL_RESULTS", "braintrust_eval_results.jsonl")
-
-
-def file_report_eval(evaluator, result: EvalResultWithSummary, verbose, jsonl):
-    try:
-        entry = {
-            "evaluator": getattr(evaluator, "name", None),
-            "summary": result.summary.as_dict(),
-        }
-        with open(OUTPUT_RESULTS_FILE, "a", encoding="utf-8") as f:
-            f.write(json.dumps(entry) + "\n")
-    except Exception as e:
-        print(f"Failed to write local results file: {e}")
-
-    # Pretty console output as well
-    results = result.results
-    failing_results = [x for x in results if x.error]
-
-    num_tests: dict[str, int] = {}
-    tests_pass_scores: dict[str, float] = {}
-    total_score = 0.0
-    total_num_tests = 0
-    for r in results:
-        category = r.metadata.get("category") if r.metadata else "unknown"
-        num_tests[category] = num_tests.get(category, 0) + 1
-        score_val = 0.0
-        if r.scores and "Tests pass" in r.scores and isinstance(r.scores["Tests pass"], (int, float)):
-            score_val = float(r.scores["Tests pass"])  # already normalized ratio per our scorer
-        tests_pass_scores[category] = tests_pass_scores.get(category, 0.0) + score_val
-        total_num_tests += 1
-        total_score += score_val
-
-    overall_rate = (total_score / total_num_tests) if total_num_tests > 0 else 0.0
-    model_name = None
-    if results and results[0].metadata and "model_name" in results[0].metadata:
-        model_name = results[0].metadata["model_name"]
-
-    print()
-    print("=== Eval Summary ===")
-    print(f"Model: {model_name if model_name else 'unknown'}")
-    print(f"Overall: {overall_rate:.2%} ({total_num_tests} tests)")
-    for category in sorted(num_tests.keys()):
-        rate = (tests_pass_scores.get(category, 0.0) / num_tests[category]) if num_tests[category] > 0 else 0.0
-        print(f"- {category}: {rate:.2%} ({num_tests[category]} tests)")
-    if failing_results:
-        print(f"Failures: {len(failing_results)} case(s)")
-    print(f"Results written to: {OUTPUT_RESULTS_FILE}")
-
-    return len(failing_results) == 0
-
-
-def file_report_run(eval_reports, verbose, jsonl):
-    return all(x for x in eval_reports)
-
-
-file_reporter = Reporter(
-    name="local-file reporter",
-    report_eval=file_report_eval,
-    report_run=file_report_run,
-)
+CONVEX_EVAL_ENDPOINT = os.getenv("CONVEX_EVAL_ENDPOINT")
+CONVEX_AUTH_TOKEN = os.getenv("CONVEX_AUTH_TOKEN")
 
 
 def convex_coding_evals(model: ModelTemplate):
